Avoid opt<size_t> with LLVMSupport
The LLVM support library that we use to parse command line options has
specializations for opt<int>, opt<unsigned> that tell the library to
parse an arbitrary numeric command line option. Using opt<size_t> seems
to fall into the enum case, so LLVMSupport expects an explicit list of
valid option values.

[REWRITTEN COMMIT: incomplete subtree]

+++ tools/graph-convert/graph-properties-convert-main.cpp
@@ -31,7 +31,7 @@ static cll::opt<galois::SourceDatabase>
                          clEnumValN(galois::SourceDatabase::MONGODB, "mongodb",
                                     "source data came from mongodb")),
              cll::init(galois::SourceDatabase::NONE));
-static cll::opt<size_t>
+static cll::opt<int>
     chunkSize("chunkSize",
               cll::desc("Chunk size for in memory arrow representation"),
               cll::init(25000));
